Dimstav23/change deployment for gdpruler and passthrough

+++ evaluation/VM/direct.sh
@@ -12,30 +12,31 @@ source $script_dir/../args_and_checks.sh
 # Call the parse_args function with your command-line arguments
 parse_args_and_checks "$@"
 
+if [[ "$server_connection" != "TCP" ]]; then
+  echo "Error: server_connection must be 'TCP' for the direct, VM experiment"
+  exit 1
+fi
+
 # compile the controller in the CVM with the appropriate encryption option
-virt-customize --add ${images_dir}/controller.img --smp $(nproc) --memsize 16384 \
+virt-customize --add ${images_dir}/gdpr.img --smp $(nproc) --memsize 16384 \
   --run-command "cd /root/GDPRuler/controller && rm -rf build && cmake -S . -B build -D CMAKE_BUILD_TYPE=Release -D ENCRYPTION_ENABLED=$encryption && cmake --build build -j$(nproc)"
 
 # GDPR controller
-results_csv_file=${script_dir}/results/passthrough_self_hosted-query_mgmt_${workload_type}-encryption_$encryption-logging_$logging.csv
-server_type="bare-metal"
+results_csv_file=${script_dir}/results/direct_CVM-query_mgmt_${workload_type}-encryption_$encryption-logging_$logging-connection_${server_connection}.csv
 for n_clients in $clients; do
   # copy the configs in the controller image
-  virt-customize --add ${images_dir}/controller.img --copy-in ${script_dir}/../configs:/root
-  # set the client config file appropriately
-  client_cfg="no_cfg"
+  virt-customize --add ${images_dir}/gdpr.img --copy-in ${script_dir}/../configs:/root
   for db in $dbs; do
     for workload in $workloads; do
       if [[ $db == "rocksdb" ]]; then
-        db_port=$rocksdb_port
-        db_address=$bare_metal_rocksdb_address
+        db_port=$direct_rocksdb_port
+        db_address=$direct_rocksdb_address
       elif [[ $db == "redis" ]]; then
-        db_port=$redis_port
-        db_address=$bare_metal_redis_address
+        db_port=$direct_redis_port
+        db_address=$direct_redis_address
       fi
-      echo "Starting a self hosted scenario run with $n_clients clients, $db store, passthrough controller, $workload and logging set to $logging"
-      run_VM_passthrough_ctl_experiment $server_type $n_clients $workload $db $db_address $db_port \
-      $controller_address $controller_port $client_cfg $results_csv_file
+      echo "Starting a direct CVM scenario run with $n_clients clients, $db store and $workload"
+      run_experiment CVM_direct $n_clients $workload $db $db_address $db_port $results_csv_file
       echo ""
     done
   done
